Remove bootstrap.pp file
This file is no longer needed as we aren't currently configuring the
system using a custom bootstrap file. Rather, the seteam/role module
contains a role::puppetmaster class that should be able to set this up
for us. Right now it's not deployed to this environment branch as the
puppetmaster is effectively externally puppetized.